requirements.txt
aiogram==3.23.0
python-dotenv==1.2.1
magic-filter==1.0.12
aiogram-sqlite-storage>=1.0.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
